requirements.txt
openai
streamlit
fastapi
uvicorn
pinecone-client
pandas
matplotlib
scikit-learn